[Develop 5주차] Item Spawner 구현완료

+++ Source/QuadLand/Gimmick/QLItemSpawner.h
@@ -18,7 +18,7 @@ public:
 	void BeginPlay();
 protected:
 	UPROPERTY(EditAnywhere, Category = Item)
-	TMap<TSubclassOf<class AQLItemBox>, int32> ItemBoxClass; //int32 는 최대 숫자를 의미함.
+	TMap<TSubclassOf<class AQLItemBox>, float> ItemBoxClass; //int32 는 최대 숫자를 의미함.
 
 	UPROPERTY(VisibleAnywhere, Category = Mesh)
 	TObjectPtr<class UStaticMeshComponent> Mesh;
